did what comments said
